refactor: move GUI and view content into view.h/.cpp
