+++ src/view.h
@@ -1,0 +1,8 @@
+#ifndef ADVANCED_IMAGE_PROCESSOR_VIEW_H__
+#define ADVANCED_IMAGE_PROCESSOR_VIEW_H__
+
+bool view_init();
+void view_loop();
+void view_cleanup();
+
+#endif // ADVANCED_IMAGE_PROCESSOR_VIEW_H__
